requirements.txt
shapely>=2.0
matplotlib>=3.7
pillow>=10.0
numpy>=1.24
svgwrite>=1.4
